Moving Script-related classes
