+++ Application/src/EvoAlg/EAController.hpp
@@ -3,7 +3,7 @@
 #include "EvoAlg/EvolutionaryAlgorithm.hpp"
 
 #include "EvoAlg/Types.hpp"
-#include "EvoAlg/Script.hpp"
+#include "EvoAlg/Script/Script.hpp"
 #include "EvoAlg/EAStatus.hpp"
 
 #include "Application/Game/GameRunner.hpp"
