+++ Application/src/EvoAlg/Script/ScriptConverter.cpp
@@ -1,7 +1,7 @@
 #include "ScriptConverter.hpp"
 
 #include "Application/Util/path.hpp"
-#include "EvoAlg/Script.hpp"
+#include "EvoAlg/Script/Script.hpp"
 
 #include "mypch.hpp"
